Separate HasDays into Class and Instances module
