+++ lib/Data/Time/Calendar.hs
@@ -4,11 +4,13 @@ module Data.Time.Calendar (
     module Data.Time.Calendar.Days,
     module Data.Time.Calendar.CalendarDiffDays,
     module Data.Time.Calendar.Gregorian,
+    module Data.Time.Calendar.HasDays,
     module Data.Time.Calendar.Week,
 ) where
 
 import Data.Time.Calendar.CalendarDiffDays
 import Data.Time.Calendar.Days
 import Data.Time.Calendar.Gregorian
+import Data.Time.Calendar.HasDays
 import Data.Time.Calendar.Week
 import Data.Time.Format ()
