--- lib/Data/Time/Calendar/Class.hs
@@ -1,9 +1,0 @@
-module Data.Time.Calendar.Class (
-    HasDays (..),
-) where
-
-import Data.Time.Calendar.Days
-
-class HasDays t where
-    firstDay :: t -> Day
-    lastDay :: t -> Day
