+++ lib/Data/Time/Calendar/HasDays.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE Safe #-}
+
+module Data.Time.Calendar.HasDays (
+    HasDays (..),
+) where
+
+import Data.Time.Calendar.HasDays.Class
+import Data.Time.Calendar.HasDays.Instances ()
